+++ HISTORY.rst
@@ -11,6 +11,10 @@ History
 
   Thanks to Quimey Vivas in `PR #346 <https://github.com/adamchainz/ec2-metadata/pull/346>`__.
 
+* Add ``domain`` and ``partition`` attributes.
+
+  Thanks to Terry Griffin in `PR #339 <https://github.com/adamchainz/ec2-metadata/pull/339>`__.
+
 2.7.0 (2022-01-10)
 ------------------